re-organize request slip components into own folder
